+++ src/main/java/com/rmaslov/blog/file/service/FileApiService.java
@@ -3,6 +3,10 @@ package com.rmaslov.blog.file.service;
 import com.mongodb.BasicDBObject;
 import com.mongodb.DBObject;
 import com.mongodb.client.gridfs.model.GridFSFile;
+import com.rmaslov.blog.auth.exceptions.AuthException;
+import com.rmaslov.blog.auth.exceptions.NotAccessException;
+import com.rmaslov.blog.auth.service.AuthService;
+import com.rmaslov.blog.base.service.CheckAccess;
 import com.rmaslov.blog.file.mapping.FileMapping;
 import com.rmaslov.blog.base.api.request.SearchRequest;
 import com.rmaslov.blog.base.api.response.SearchResponse;
@@ -11,6 +15,7 @@ import com.rmaslov.blog.file.exception.FileNotExistException;
 import com.rmaslov.blog.file.model.FileDoc;
 import com.rmaslov.blog.file.repository.FileRepository;
 import com.rmaslov.blog.user.exception.UserNotExistException;
+import com.rmaslov.blog.user.model.UserDoc;
 import com.rmaslov.blog.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.bson.types.ObjectId;
@@ -30,15 +35,16 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-public class FileApiService {
+public class FileApiService extends CheckAccess<FileDoc> {
     private final FileRepository fileRepository;
     private final MongoTemplate mongoTemplate;
     private final GridFsTemplate gridFsTemplate;
     private final GridFsOperations operations;
-    private final UserRepository userRepository;
+    private final AuthService authService;
 
-    public FileDoc create(MultipartFile file, ObjectId ownerId) throws IOException, UserNotExistException {
-        if(userRepository.findById(ownerId).isEmpty()) throw new UserNotExistException();
+
+    public FileDoc create(MultipartFile file) throws IOException, AuthException {
+        UserDoc owner = authService.currentUser();
 
         DBObject metaData = new BasicDBObject();
         metaData.put("type", file.getContentType());
@@ -51,7 +57,7 @@ public class FileApiService {
         FileDoc fileDoc = FileDoc.builder()
                 .id(id)
                 .title(file.getOriginalFilename())
-                .ownerId(ownerId)
+                .ownerId(owner.getId())
                 .contentType(file.getContentType())
                 .build();
 
@@ -89,8 +95,19 @@ public class FileApiService {
         return SearchResponse.of(fileDocs, count);
     }
 
-    public void delete(ObjectId id){
+    public void delete(ObjectId id) throws AuthException, NotAccessException, ChangeSetPersister.NotFoundException {
+        checkAccess(fileRepository.findById(id).orElseThrow(ChangeSetPersister.NotFoundException::new));
         gridFsTemplate.delete(new Query(Criteria.where("_id").is(id)));
         fileRepository.deleteById(id);
+    }
+
+    @Override
+    protected ObjectId getOwnerFromEntity(FileDoc entity) {
+        return entity.getOwnerId();
+    }
+
+    @Override
+    protected AuthService authService() {
+        return authService;
     }
 }
